binary tree test
